#44 add to favorite or remove if exist Small changes

+++ FastAPI/favorite/crud.py
@@ -1,4 +1,5 @@
 from typing import Optional
+from typing import Dict, List, Mapping
 
 from FastAPI.config import database
 
@@ -9,13 +10,13 @@ async def get_user_id(token: str) -> Optional[int]:
     return row["user_id"] if row else None
 
 
-async def offers_in_favorite(user_id: int):
+async def get_favorite_offers_of_user(user_id: int) -> List[Mapping]:
     query = """
     SELECT
         cover_image,
         currency,
         description,
-        id,
+        offers_offer.id,
         is_deliverable,
         price,
         pub_date,
@@ -36,8 +37,7 @@ async def add_offer_to_favorite_or_delete(user_id: int, offer_id: str) -> None:
     WHERE offer_id=:offer_id
       AND user_id=:user_id
     """
-    values = {"user_id": user_id, "offer_id": offer_id}
-    row = await database.fetch_one(query=query, values=values)
+    row = await database.fetch_one(query=query, values={"user_id": user_id, "offer_id": offer_id})
     if row:
         await remove_favorite(user_id, offer_id)
     else:
@@ -45,21 +45,14 @@ async def add_offer_to_favorite_or_delete(user_id: int, offer_id: str) -> None:
 
 
 async def remove_favorite(user_id: int, offer_id: str) -> None:
-    query = """
-    DELETE FROM offers_offer_favorite
-    WHERE offer_id =:offer_id
-    AND user_id =:user_id
-    """
+    query = "DELETE FROM offers_offer_favorite WHERE offer_id =:offer_id AND user_id =:user_id"
     await database.execute(
         query=query, values={"user_id": user_id, "offer_id": offer_id}
     )
 
 
 async def add_favorite(user_id: int, offer_id: str) -> None:
-    query = """
-    INSERT INTO offers_offer_favorite (user_id, offer_id)
-    VALUES (:user_id, :offer_id)
-    """
+    query = "INSERT INTO offers_offer_favorite (user_id, offer_id) VALUES (:user_id, :offer_id)"
     await database.execute(
         query=query, values={"user_id": user_id, "offer_id": offer_id}
     )
